requirements.txt
--index-url https://pypi.drwholdings.com
#requests==2.18.4
flask-orator==0.2.0
cleo==0.6 # for flask-orator compatibility
#pkg-resources==0.0.0
PyYAML==5.3.1
Flask==0.12.2
ipython==6.4.0
